# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "App-Api",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/projects": {
            "get": {
                "tags": [
                    "Projects"
                ],
                "summary": "Get all projects",
                "operationId": "fbc21feeb3df8a2cb519efb6480ae7d9",
                "responses": {
                    "200": {
                        "description": "A list of all projects"
                    }
                }
            },
            "post": {
                "tags": [
                    "Projects"
                ],
                "summary": "Create a new project",
                "operationId": "a9d5bd5914cf75eaebafcf8b9e320dcf",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "userId",
                                    "backImg"
                                ],
                                "properties": {
                                    "name": {
                                        "description": "The name of the project",
                                        "type": "string",
                                        "example": "My Project"
                                    },
                                    "userId": {
                                        "description": "The ID of the user associated with the project",
                                        "type": "integer",
                                        "example": 123
                                    },
                                    "backImg": {
                                        "description": "The URL or file path of the project's background image",
                                        "type": "string",
                                        "example": "https://example.com/project-image.jpg"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "The created project"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
        "/api/projects/{id}": {
            "get": {
                "tags": [
                    "Projects"
                ],
                "summary": "Get a project by ID",
                "operationId": "865085b82af27d647cf1f2a772aa7927",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "The ID of the project",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "The project data"
                    },
                    "404": {
                        "description": "Project not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Projects"
                ],
                "summary": "Update an existing project",
                "operationId": "fe600b7ffce78456bce67c51b9eca51d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "The ID of the project to update",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "The updated project"
                    },
                    "404": {
                        "description": "Project not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Projects"
                ],
                "summary": "Delete a project",
                "operationId": "d81ac2b0b51fa7143c69f1a5f4c27eb0",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "The ID of the project to delete",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Project successfully deleted"
                    },
                    "404": {
                        "description": "Project not found"
                    }
                }
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}